Created expectations.
This is the source code the 'expect' and a sample rule set titled the 'triswapnot' machine.  It's designed to be a game.  The goal of the game is to create the longest sentance with the shortest (DEBUG) output.  This is obviously important that every player is running the exact same code.  This is the first reference version.  V1.