Trying to decide if i should change 'swap' to 'swch'
Swap was difficult to hear for my father over the phone and it lead to confusion so it led me to belive that it might not be the correct *sound* for the concept.  A look in the dictionary gave me switch, but the 'it' makes the word too long and stand out from the other 2 and not roll off the tounge as easily as swap.  This led me to /abbreviate/, something i normally hate in coding as it requires energy from the brain to decde the abbreviation, something that should be minimized.  But I figure with so few words (3) that an abbreviation for one of the words won't be  a problem.

TODO: update documentation and code to change swap to swtch